Try and add version flag
